+++ activity/jd_global.js
@@ -1,34 +1,34 @@
 /*
 环球挑战赛
-活动时间：2021-02-02 至 2021-02-22
-1个号可以助力5人，需要5人助力，如每天跑满能换1000+京豆
+活动时间：2021-03-08 至 2021-03-31
 多个账号会相互互助
-活动地址：https://gmart.jd.com/?appId=54935130
+活动地址：https://ghttps://gmart.jd.com/?appId=54935130mart.jd.com/?appId=54935130
 活动入口：京东app搜索京东国际-环球挑战赛
 已支持IOS双京东账号,Node.js支持N个京东账号
 脚本兼容: QuantumultX, Surge, Loon, JSBox, Node.js
 ============Quantumultx===============
 [task_local]
 #环球挑战赛
-0 9,12,20,21 2-22 2 * https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js, tag=环球挑战赛, img-url=https://raw.githubusercontent.com/yogayyy/Scripts/main/Icon/lxk0301/jd_global.png, enabled=true
+0 9,12,20,21 8-31 3 * https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js, tag=环球挑战赛, img-url=https://raw.githubusercontent.com/Orz-3/mini/master/Color/jd.png, enabled=true
 
 ================Loon==============
 [Script]
-cron "0 9,12,20,21 2-22 2 *" script-path=https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js,tag=环球挑战赛
+cron "0 9,12,20,21 8-31 3 *" script-path=https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js,tag=环球挑战赛
 
 ===============Surge=================
-环球挑战赛 = type=cron,cronexp="0 9,12,20,21 2-22 2 *",wake-system=1,timeout=3600,script-path=https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js
+环球挑战赛 = type=cron,cronexp="0 9,12,20,21 8-31 3 *",wake-system=1,timeout=3600,script-path=https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js
 
 ============小火箭=========
-环球挑战赛 = type=cron,script-path=https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js, cronexpr="0 9,12,20,21 2-22 2 *", timeout=3600, enable=true
+环球挑战赛 = type=cron,script-path=https://gitee.com/lxk0301/jd_scripts/raw/master/jd_global.js, cronexpr="0 9,12,20,21 8-31 3 *", timeout=3600, enable=true
  */
 const $ = new Env('环球挑战赛');
+!function(n){"use strict";function r(n,r){var t=(65535&n)+(65535&r);return(n>>16)+(r>>16)+(t>>16)<<16|65535&t}function t(n,r){return n<<r|n>>>32-r}function u(n,u,e,o,c,f){return r(t(r(r(u,n),r(o,f)),c),e)}function e(n,r,t,e,o,c,f){return u(r&t|~r&e,n,r,o,c,f)}function o(n,r,t,e,o,c,f){return u(r&e|t&~e,n,r,o,c,f)}function c(n,r,t,e,o,c,f){return u(r^t^e,n,r,o,c,f)}function f(n,r,t,e,o,c,f){return u(t^(r|~e),n,r,o,c,f)}function i(n,t){n[t>>5]|=128<<t%32,n[14+(t+64>>>9<<4)]=t;var u,i,a,h,g,l=1732584193,d=-271733879,v=-1732584194,C=271733878;for(u=0;u<n.length;u+=16)i=l,a=d,h=v,g=C,d=f(d=f(d=f(d=f(d=c(d=c(d=c(d=c(d=o(d=o(d=o(d=o(d=e(d=e(d=e(d=e(d,v=e(v,C=e(C,l=e(l,d,v,C,n[u],7,-680876936),d,v,n[u+1],12,-389564586),l,d,n[u+2],17,606105819),C,l,n[u+3],22,-1044525330),v=e(v,C=e(C,l=e(l,d,v,C,n[u+4],7,-176418897),d,v,n[u+5],12,1200080426),l,d,n[u+6],17,-1473231341),C,l,n[u+7],22,-45705983),v=e(v,C=e(C,l=e(l,d,v,C,n[u+8],7,1770035416),d,v,n[u+9],12,-1958414417),l,d,n[u+10],17,-42063),C,l,n[u+11],22,-1990404162),v=e(v,C=e(C,l=e(l,d,v,C,n[u+12],7,1804603682),d,v,n[u+13],12,-40341101),l,d,n[u+14],17,-1502002290),C,l,n[u+15],22,1236535329),v=o(v,C=o(C,l=o(l,d,v,C,n[u+1],5,-165796510),d,v,n[u+6],9,-1069501632),l,d,n[u+11],14,643717713),C,l,n[u],20,-373897302),v=o(v,C=o(C,l=o(l,d,v,C,n[u+5],5,-701558691),d,v,n[u+10],9,38016083),l,d,n[u+15],14,-660478335),C,l,n[u+4],20,-405537848),v=o(v,C=o(C,l=o(l,d,v,C,n[u+9],5,568446438),d,v,n[u+14],9,-1019803690),l,d,n[u+3],14,-187363961),C,l,n[u+8],20,1163531501),v=o(v,C=o(C,l=o(l,d,v,C,n[u+13],5,-1444681467),d,v,n[u+2],9,-51403784),l,d,n[u+7],14,1735328473),C,l,n[u+12],20,-1926607734),v=c(v,C=c(C,l=c(l,d,v,C,n[u+5],4,-378558),d,v,n[u+8],11,-2022574463),l,d,n[u+11],16,1839030562),C,l,n[u+14],23,-35309556),v=c(v,C=c(C,l=c(l,d,v,C,n[u+1],4,-1530992060),d,v,n[u+4],11,1272893353),l,d,n[u+7],16,-155497632),C,l,n[u+10],23,-1094730640),v=c(v,C=c(C,l=c(l,d,v,C,n[u+13],4,681279174),d,v,n[u],11,-358537222),l,d,n[u+3],16,-722521979),C,l,n[u+6],23,76029189),v=c(v,C=c(C,l=c(l,d,v,C,n[u+9],4,-640364487),d,v,n[u+12],11,-421815835),l,d,n[u+15],16,530742520),C,l,n[u+2],23,-995338651),v=f(v,C=f(C,l=f(l,d,v,C,n[u],6,-198630844),d,v,n[u+7],10,1126891415),l,d,n[u+14],15,-1416354905),C,l,n[u+5],21,-57434055),v=f(v,C=f(C,l=f(l,d,v,C,n[u+12],6,1700485571),d,v,n[u+3],10,-1894986606),l,d,n[u+10],15,-1051523),C,l,n[u+1],21,-2054922799),v=f(v,C=f(C,l=f(l,d,v,C,n[u+8],6,1873313359),d,v,n[u+15],10,-30611744),l,d,n[u+6],15,-1560198380),C,l,n[u+13],21,1309151649),v=f(v,C=f(C,l=f(l,d,v,C,n[u+4],6,-145523070),d,v,n[u+11],10,-1120210379),l,d,n[u+2],15,718787259),C,l,n[u+9],21,-343485551),l=r(l,i),d=r(d,a),v=r(v,h),C=r(C,g);return[l,d,v,C]}function a(n){var r,t="",u=32*n.length;for(r=0;r<u;r+=8)t+=String.fromCharCode(n[r>>5]>>>r%32&255);return t}function h(n){var r,t=[];for(t[(n.length>>2)-1]=void 0,r=0;r<t.length;r+=1)t[r]=0;var u=8*n.length;for(r=0;r<u;r+=8)t[r>>5]|=(255&n.charCodeAt(r/8))<<r%32;return t}function g(n){return a(i(h(n),8*n.length))}function l(n,r){var t,u,e=h(n),o=[],c=[];for(o[15]=c[15]=void 0,e.length>16&&(e=i(e,8*n.length)),t=0;t<16;t+=1)o[t]=909522486^e[t],c[t]=1549556828^e[t];return u=i(o.concat(h(r)),512+8*r.length),a(i(c.concat(u),640))}function d(n){var r,t,u="";for(t=0;t<n.length;t+=1)r=n.charCodeAt(t),u+="0123456789abcdef".charAt(r>>>4&15)+"0123456789abcdef".charAt(15&r);return u}function v(n){return unescape(encodeURIComponent(n))}function C(n){return g(v(n))}function A(n){return d(C(n))}function m(n,r){return l(v(n),v(r))}function s(n,r){return d(m(n,r))}function b(n,r,t){return r?t?m(r,n):s(r,n):t?C(n):A(n)}$.md5=b}();
 
 const notify = $.isNode() ? require('./sendNotify') : '';
 //Node.js用户请在jdCookie.js处填写京东ck;
 const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
 let jdNotify = true;//是否关闭通知，false打开通知推送，true关闭通知推送
-const randomCount = 0 ;
+const randomCount =0 ;
 //IOS等用户直接用NobyDa的jd cookie
 
 let cookiesArr = [], cookie = '', message;
@@ -39,19 +39,13 @@ if ($.isNode()) {
   if (process.env.JD_DEBUG && process.env.JD_DEBUG === 'false') console.log = () => {
   };
 } else {
-  let cookiesData = $.getdata('CookiesJD') || "[]";
-  cookiesData = jsonParse(cookiesData);
-  cookiesArr = cookiesData.map(item => item.cookie);
-  cookiesArr.reverse();
-  cookiesArr.push(...[$.getdata('CookieJD2'), $.getdata('CookieJD')]);
-  cookiesArr.reverse();
-  cookiesArr = cookiesArr.filter(item => item !== "" && item !== null && item !== undefined);
+  cookiesArr = [$.getdata('CookieJD'), $.getdata('CookieJD2'), ...jsonParse($.getdata('CookiesJD') || "[]").map(item => item.cookie)].filter(item => !!item);
 }
 
 const JD_API_HOST = 'https://api.m.jd.com/', actCode = 'visa-card-001';
 const inviteCodes = [
-  'S1J3TVNRc0tzRExxU011OUoxRU8zQT09@aklCQnZUV2hYcktrOG1CTVlRTW5DbmRQSWhIVUVEajZtaGFnRXdCc2dGQT0=@c2g0M3gvVmVCZjI3RHNFNXB1VCtzdGlBeGFnUUlZQU85ejRKQW54UG9LUT0=@R2dhRUd6OFNBbnE5MERra2FHR01TQWZaeDF6Y0laaXQ4ZENwNmt5YmRSRT0=@QWxxaSt0UEFNTmFVbFNYeXJKMDhYS2YvOTZyT3pDbmRZcm1lWjJHWklHYz0=@NDRZcGpEcHprR2dyeXFxRjBIbCt0UT09@anFXc3FzdzJUeE9GM3VOMFhzcElxbjFkQVhqVW5VSDNLZzZaUHNVUkxVcz0=',
-  'S1J3TVNRc0tzRExxU011OUoxRU8zQT09@aklCQnZUV2hYcktrOG1CTVlRTW5DbmRQSWhIVUVEajZtaGFnRXdCc2dGQT0=@c2g0M3gvVmVCZjI3RHNFNXB1VCtzdGlBeGFnUUlZQU85ejRKQW54UG9LUT0=@R2dhRUd6OFNBbnE5MERra2FHR01TQWZaeDF6Y0laaXQ4ZENwNmt5YmRSRT0=@QWxxaSt0UEFNTmFVbFNYeXJKMDhYS2YvOTZyT3pDbmRZcm1lWjJHWklHYz0=@NDRZcGpEcHprR2dyeXFxRjBIbCt0UT09@anFXc3FzdzJUeE9GM3VOMFhzcElxbjFkQVhqVW5VSDNLZzZaUHNVUkxVcz0=',
+  'S1J3TVNRc0tzRExxU011OUoxRU8zQT09@aklCQnZUV2hYcktrOG1CTVlRTW5DbmRQSWhIVUVEajZtaGFnRXdCc2dGQT0=@R2dhRUd6OFNBbnE5MERra2FHR01TQWZaeDF6Y0laaXQ4ZENwNmt5YmRSRT0=@QWxxaSt0UEFNTmFVbFNYeXJKMDhYS2YvOTZyT3pDbmRZcm1lWjJHWklHYz0=@c2g0M3gvVmVCZjI3RHNFNXB1VCtzdGlBeGFnUUlZQU85ejRKQW54UG9LUT0=@NDRZcGpEcHprR2dyeXFxRjBIbCt0UT09@WUtWSzh1WTR2THhWdHBYZHRCZW5IUT09@RnA2dUpMRFVSMHU2emsvNm40bHc5VCt2MzhtUjViV0g1ZVZQL3NkaHBDQT0=@anFXc3FzdzJUeE9GM3VOMFhzcElxbjFkQVhqVW5VSDNLZzZaUHNVUkxVcz0=@L2lObkpxWVQxNlViZTJnNkt1NlN3VVBzT1pYZW0wekxqbTdIaTh6eFNlUT0=@c28veGFualpnTzg2WHM5bnhkdFUyM2hlNkNHMHluNWpDYll0S21xYjk0QT0=@cUZ5RUMyZ3VyaERLdXNQS1NzUUtVdz09@Yy9KaUYxYTVUWkIzVjlISHJJWTRiUT09@MXBBZFZJYXVQUkZQbTZzQnZzc2pPL1AxdjdzMjRURTBIcm5MUWtDQXlMdz0=@VWZUL3dMVTlVbG9DVW1CZmkzRU1TaEU4aEQ1alVpUHQxb0paSmtYUlM1ST0=',
+  'S1J3TVNRc0tzRExxU011OUoxRU8zQT09@aklCQnZUV2hYcktrOG1CTVlRTW5DbmRQSWhIVUVEajZtaGFnRXdCc2dGQT0=@R2dhRUd6OFNBbnE5MERra2FHR01TQWZaeDF6Y0laaXQ4ZENwNmt5YmRSRT0=@QWxxaSt0UEFNTmFVbFNYeXJKMDhYS2YvOTZyT3pDbmRZcm1lWjJHWklHYz0=@c2g0M3gvVmVCZjI3RHNFNXB1VCtzdGlBeGFnUUlZQU85ejRKQW54UG9LUT0=@NDRZcGpEcHprR2dyeXFxRjBIbCt0UT09@WUtWSzh1WTR2THhWdHBYZHRCZW5IUT09@RnA2dUpMRFVSMHU2emsvNm40bHc5VCt2MzhtUjViV0g1ZVZQL3NkaHBDQT0=@anFXc3FzdzJUeE9GM3VOMFhzcElxbjFkQVhqVW5VSDNLZzZaUHNVUkxVcz0=@L2lObkpxWVQxNlViZTJnNkt1NlN3VVBzT1pYZW0wekxqbTdIaTh6eFNlUT0=@c28veGFualpnTzg2WHM5bnhkdFUyM2hlNkNHMHluNWpDYll0S21xYjk0QT0=@cUZ5RUMyZ3VyaERLdXNQS1NzUUtVdz09@Yy9KaUYxYTVUWkIzVjlISHJJWTRiUT09@MXBBZFZJYXVQUkZQbTZzQnZzc2pPL1AxdjdzMjRURTBIcm5MUWtDQXlMdz0=@VWZUL3dMVTlVbG9DVW1CZmkzRU1TaEU4aEQ1alVpUHQxb0paSmtYUlM1ST0=',
 ];
 $.invites = [];
 !(async () => {
@@ -82,18 +76,7 @@ $.invites = [];
       await jdGlobal()
     }
   }
-  for (let i = 0; i < cookiesArr.length; i++) {
-    if (cookiesArr[i]) {
-      cookie = cookiesArr[i];
-      $.canHelp = true;
-      for (let code of $.invites) {
-        if (!code) continue
-        await helpFriend(code)
-        if(!$.canHelp) break
-        await $.wait(1000)
-      }
-    }
-  }
+
 })()
   .catch((e) => {
     $.log('', `❌ ${$.name}, 失败! 原因: ${e}!`, '')
@@ -120,6 +103,7 @@ async function jdGlobal() {
 async function helpFriends() {
   $.canHelp = true
   for (let code of $.newShareCodes) {
+    console.log(`去助力好友${code}`)
     if (!code) continue
     await helpFriend(code)
     if(!$.canHelp) break
@@ -132,6 +116,9 @@ function showMsg() {
   return new Promise(resolve => {
     message += `本次运行获得${$.earn}里程，${$.beans}京豆，共计${$.score}里程`
     $.msg($.name, '', `京东账号${$.index}${$.nickName}\n${message}`);
+	if($.beans>0){
+		notify.sendNotify(`${$.name}`, `账号${$.index} - ${$.nickName || $.UserName}\n${message}\n`);
+		}
     resolve()
   })
 }
@@ -198,7 +185,8 @@ async function getTask() {
               let task = [...timeLimitTask, ...commonTask]
               for (let vo of task) {
                 if (vo['taskName'] === '每日邀请好友') {
-                  console.log(`您的好友助力码为 ${vo['jingCommand']['keyOpenapp'].match(/masterPin":"(.*)","/)[1]}`)
+                  // console.log(`您的好友助力码为 ${vo['jingCommand']['keyOpenapp'].match(/masterPin":"(.*)","/)[1]}`)
+                  console.log(`\n【京东账号${$.index}（${$.nickName || $.UserName}）的${$.name}好友互助码】${vo['jingCommand']['keyOpenapp'].match(/masterPin":"(.*)","/)[1]}\n`);
                   $.invites.push(vo['jingCommand']['keyOpenapp'].match(/masterPin":"(.*)","/)[1]);
                 }
                 if (['70', '50', '30', '40'].includes(vo['taskType'])) {
@@ -260,7 +248,7 @@ async function helpFriend(inviterPin) {
   return new Promise(resolve => {
     $.get(taskUrl("inviteHelp", {
       "inviterPin": inviterPin,
-      "taskId": "3",
+      "taskId": "51",
       "pageType": "doHelp",
       "headImg": "",
       "username": "",
@@ -393,6 +381,14 @@ function requireConfig() {
 }
 
 function taskUrl(function_id, body = {}) {
+  function getSign(data) {
+    let t = +new Date()
+
+    return {sealsTs: t, seals: $.md5(`${data.taskId}${data.inviterPin?data.inviterPin:''}${t}Ea6YXT`)}
+  }
+  if(body['taskId']) {
+    body = {...body, ...getSign(body)}
+  }
   return {
     url: `${JD_API_HOST}/client.action?functionId=${function_id}&body=${escape(JSON.stringify(body))}&appid=global_mart&time=${new Date().getTime()}`,
     headers: {
@@ -450,7 +446,11 @@ function TotalBean() {
               $.isLogin = false; //cookie过期
               return
             }
-            $.nickName = data['base'].nickname;
+            if (data['retcode'] === 0) {
+              $.nickName = (data['base'] && data['base'].nickname) || $.UserName;
+            } else {
+              $.nickName = $.UserName
+            }
           } else {
             console.log(`京东服务器返回空数据`)
           }
